# openapi.yaml
...
openapi: "3.0.3"
info:
  title: sc-agent
  description: API to interact with sc-agent
  version: 1.0.0
  contact:
    name: soeren schneider
    url: https://github.com/soerenschneider/sc-agent
    email: "56670304+soerenschneider@users.noreply.github.com" # this is just for linting reasons
  license:
    name: "NONE" # this is just for linting reasons
    url: "none" # this is just for linting reasons
servers:
  - url: http://localhost:9999
    description: Local server
...
paths:
...
  /v1/certs/ssh:
    get:
      operationId: certsSshGetCertificates
      summary: Get the configuration of all managed ssh certificates
      description: >
        Returns a list of all configured managed ssh certificates.
      tags:
        - certs
        - ssh
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - user
              - host
          example: "user"
          description: Specifies whether to return cert of type user or host. See Vault documentation.
      responses:
        '200':
          description: Object that contains the list of all configured certs
          content:
            application/json:
              schema:
                type: array
                x-go-type-skip-optional-pointer: true
                items:
                  $ref: '#/components/schemas/SshManagedCertificatesList'
                description: Array holding conf
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/certs/ssh/{id}:
    get:
      operationId: certsSshGetCertificate
      summary: Get the configuration of a single managed ssh certificate
      description: >
        Returns the configuration of a managed ssh certificates for a given id. The parameter id is supplied via path.
      tags:
        - certs
        - ssh
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "my-host-cert"
          description: The id of the SSH certificate you want to retrieve information for
      responses:
        '200':
          description: Configuration of the managed ssh certificate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SshManagedCertificate"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/certs/ssh/issue-requests:
    post:
      operationId: certsSshPostIssueRequests
      summary: Sign a SSH public key
      description: >
        Signs a SSH public key and writes the signature to a file. The appropriate public key is picked from the configuration and is passed via path parameter. The public key is only signed if 
        a signature does not exist already or if the configured threshold of the remaining lifetime is exceeded.
      tags:
        - certs
        - ssh
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          example: "my-user-cert"
          description: The id of the SSH host certificate you want to sign
        - name: force-renewal
          in: query
          required: false
          schema:
            type: boolean
          description: Do not check the lifetime of the existing certificate against the threshold and force a new signature of the host key
      responses:
        '200':
          description: Certificate already exists and is still valid, no new signature was needed
        '201':
          description: New signature created, either because certificate did not exist, was expired, was below the lifetime threshold or because force_renewal was set to true.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/power-state/conditional-reboot:
    put:
      operationId: powerConditionalRebootPostStatus
      summary: Unpause reboot status
      description: Set the pause status of the conditional reboot service
      tags:
        - power
      parameters:
        - name: "action"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - "pause"
              - "unpause"
          example: "pause"
          description: Action to perform for the conditional-reboot service
      responses:
        '200':
          description: Reboot status unpaused
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/power-state/conditional-reboot/status:
    get:
      operationId: powerConditionalRebootGetStatus
      summary: Get reboot status
      description: Return the current reboot status.
      tags:
        - power
      responses:
        '200':
          description: Reboot status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: paused
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/power-state:
    post:
      operationId: powerPostAction
      summary: Directly interact with the power status of the system
      description: >
        Send an action to the local system that either tries to shutdown or reboot the machine this API is running on.
        The appropriate action is set via path variable.
      tags:
        - power
      parameters:
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum:
              - reboot
              - shutdown
          example: reboot
          description: Action to perform for the machine
      responses:
        '200':
          description: Machine shutdown successful
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/replication/secrets/sync-requests:
    post:
      operationId: replicationPostSecretsRequests
      summary: Replicate a secret from Vault
      description: >
        Replicates a predefined secret from Vault to the local filesystem. The secret config is defined by the secret path which is expected as query parameter.
      tags:
        - secrets
      parameters:
        - name: secret-path
          in: query
          required: true
          schema:
            type: string
          example: "prod/database"
          description: The KV2 path of the secret that you want to trigger the sync for
      responses:
        '200':
          description: Successfully synced secret
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/replication/secrets/items:
    get:
      operationId: replicationGetSecretsItemsList
      summary: "Returns current configuration"
      description: Returns the current configuration of the secret replication component.
      tags:
        - secrets
      responses:
        '200':
          description: The current configuration of the secret syncer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplicationSecretsItemsList"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/replication/secrets/items/{id}:
    get:
      operationId: replicationGetSecretsItem
      summary: "Returns the replication status of a single item"
      description: >
        Returns the replication status of a single secret replication item. The id of the item is passed via path
        variable.
      tags:
        - secrets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "prod/db"
          description: The id which translates to the secret path of the to-be-synced item in Vault.
      responses:
        '200':
          description: Object containing the status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplicationSecretsItem"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/services/{unit}/status:
    put:
      operationId: servicesUnitStatusPut
      summary: Interact with a system service
      description: Interacts with a system service. Unit and action are set via path.
      tags:
        - services
      parameters:
        - name: unit
          in: path
          required: true
          schema:
            type: string
          description: The name of the system service unit to be restarted
          example: sshd.service
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum:
              - restart
              - stop
              - start
          example: start
          description: Action to perform for the system service unit
      responses:
        '200':
          description: Action carried out successfully to system service unit
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/services/{unit}/logs:
    get:
      operationId: servicesUnitLogsGet
      summary: Get logs of a system service unit
      description: Retrieve logs for a system unit. The unit must be provided in the path.
      tags:
        - services
      parameters:
        - name: unit
          description: The name of the system service unit to request logs for
          in: path
          required: true
          schema:
            type: string
          example: "sshd.service"
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceLogsData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
components:
  schemas:
    Problem:
      description: A generic HTTP problem
      xml:
        name: problem
        namespace: urn:ietf:rfc:7807
      properties:
        type:
          type: "string"
          description: "A URI reference that identifies the problem type"
          example: "https://example.com/probs/out-of-credit"
        title:
          type: "string"
          description: "A short, human-readable summary of the problem type"
          example: "You do not have enough credit."
        detail:
          type: "string"
          description: "A human-readable explanation specific to this occurrence of the problem"
          example: "Your current balance is 30, but that costs 50."
        instance:
          type: "string"
          description: "A URI reference that identifies the specific occurrence of the problem"
          example: "/account/12345/msgs/abc"
        status:
          type: "integer"
          description: "The HTTP status code"
          example: 400
      example:
        type: "https://example.com/probs/out-of-credit"
        title: "You do not have enough credit."
        detail: "You do not have enough credit."
        instance: "/account/12345/msgs/abc"
        status: 400
...
    ReplicationSecretsItemsList:
      type: object
      title: SecretReplicationItems
      description: "The configuration of all configured managed secrets replication items"
      properties:
        data:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/ReplicationSecretsItem'
          description: The sync items
      example:
        data:
          - secret_path: "prod/db"
            formatter: "json"
            dest_uri: "/tmp/db.json"
    ReplicationSecretsItem:
      type: object
      title: SecretReplicationItem
      description: "Configuration and status of a single secret replication item"
      properties:
        id:
          type: string
          example: "database"
          x-go-type-skip-optional-pointer: true
          description: id of the item
        secret_path:
          type: string
          example: "prod/db"
          x-go-type-skip-optional-pointer: true
          description: path of the secret to read and sync to the local filesystem
        formatter:
          type: string
          example: "json"
          x-go-type-skip-optional-pointer: true
          description: name of the formatter
        dest_uri:
          type: string
          example: "/tmp/db.json"
          x-go-type-skip-optional-pointer: true
          description: destination path where the read secret should be writen to
        status:
          type: string
          example: "synced"
          description: the status of the synced secret
          enum:
            - unknown
            - synced
            - failed
      example:
        secret_path: "prod/db"
        formatter: "json"
        dest_uri: "/tmp/db.json"
        status: "failed"
...
    ServiceLogsData:
      type: object
      description: Logs data for a service
      properties:
        data:
          $ref: '#/components/schemas/ServiceLogs'
        error:
          type: string
          nullable: true
          example: null
      example:
        data:
          logs:
            - "Feb 03 21:48:52 localhost.localdomain kernel: SELinux: 2048 avtab hash slots, 104131 rules."
            - "Feb 03 21:48:52 localhost.localdomain kernel: SELinux: 2048 avtab hash slots, 104131 rules."
    ServiceLogs:
      type: object
      title: Logs
      description: Represents log data. Each entry in the array represents a single line of data.
      properties:
        logs:
          type: array
          x-go-type-skip-optional-pointer: true
          maxItems: 1000
          items:
            type: string
      example:
        logs:
          - "Feb 03 21:48:52 localhost.localdomain kernel: SELinux: 2048 avtab hash slots, 104131 rules."
          - "Feb 03 21:48:52 localhost.localdomain kernel: SELinux: 2048 avtab hash slots, 104131 rules."
...
    SshManagedCertificate:
      type: object
      title: SshManagedCertificate
      description: Represents the configuration and storage information of a managed SSH certificate
      properties:
        certificate_config:
          $ref: '#/components/schemas/SshCertificateConfig'
        storage_config:
          $ref: '#/components/schemas/SshCertificateStorage'
        certificate:
          $ref: '#/components/schemas/SshCertificateData'
    SshManagedCertificatesList:
      type: object
      title: SshManagedCertificatesList
      properties:
        data:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/SshManagedCertificate'
      description: Returns a list of all the configured managed SSH certificate
    SshCertificateStorage:
      type: object
      title: SshCertificateStorage
      description: The storage configuration of a managed SSH certificate
      properties:
        public_key_file:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The file that contains the public key to be signed
        certificate_file:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The file that the certificate will be written to
...
    SshCertificateConfig:
      type: object
      title: SshCertificateConfig
      description: Represents the configuration of a managed SSH certificate
      properties:
        id:
          type: "string"
          x-go-type-skip-optional-pointer: true
          description: "The id of this managed certificate"
          example: "user-cert"
        role:
          type: "string"
          x-go-type-skip-optional-pointer: true
          description: "Name of the vault role"
          example: "my-role"
        principals:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            type: "string"
            description: "Valid principals for the signed cert"
            example: "admin"
          example: ["admin", "nimda"]
        public_key_file:
          type: "string"
          x-go-type-skip-optional-pointer: true
          description: "The file to read the SSH public key from"
          example: "/home/user/.ssh/id_rsa.pub"
        certificate_file:
          type: "string"
          x-go-type-skip-optional-pointer: true
          description: "The file to write the signed cert to"
          example: "/home/user/.ssh/id_rsa.crt"
        ttl:
          type: "string"
          x-go-type-skip-optional-pointer: true
          description: "The desired TTL of the signature"
          example: "48h"
        cert_type:
          x-go-type-skip-optional-pointer: true
          type: string
      example:
        id: "user-cert"
        role: "my-role"
        principals: ["admin", "nimda"]
        publicKeyFile: "/home/user/.ssh/id_rsa.pub"
        certificateFile: "/home/user/.ssh/id_rsa.crt"
        ttl: "48h"
        certType: "rsa"
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden for the provided credentials.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden for the provided credentials.
    InternalServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
    NotImplemented:
      description: Not Implemented
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-implemented
            title: Not Implemented
            status: 501
            detail: This function is not enabled.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-implemented
            title: Not Implemented
            status: 501
            detail: This function is not enabled.
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: You do not have the necessary permissions.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: You do not have the necessary permissions.